Add compat.py module
