+++ src/estimagic/utilities.py
@@ -6,7 +6,6 @@ import cloudpickle
 import numpy as np
 import pandas as pd
 from scipy.linalg import ldl, qr
-from estimagic.config import IS_PANDAS_VERSION_NEWER_OR_EQUAL_TO_2_1_0
 
 with warnings.catch_warnings():
     warnings.simplefilter("ignore", category=UserWarning)
@@ -322,28 +321,3 @@ def get_rng(seed):
     else:
         raise TypeError("seed type must be in {None, int, numpy.random.Generator}.")
     return rng
-
-
-def pandas_df_map(df, func, na_action=None, **kwargs):
-    """Apply a function to a Dataframe elementwise.
-
-    pandas has depricated the .applymap() function with version 2.1.0. This function
-    calls either .map() (if pandas version is greater or equal to 2.1.0) or .applymap()
-    (if pandas version is smaller than 2.1.0).
-
-    Args:
-        df (pd.DataFrame): A pandas DataFrame.
-        func (callable): Python function, returns a single value from a single value.
-        na_action (str): If 'ignore', propagate NaN values, without passing them to
-            func. If None, pass NaN values to func. Default is None.
-        **kwargs: Additional keyword arguments to pass as keywords arguments to func.
-
-    Returns:
-        pd.DataFrame: Transformed DataFrame.
-
-    """
-    if IS_PANDAS_VERSION_NEWER_OR_EQUAL_TO_2_1_0:
-        out = df.map(func, na_action=na_action, **kwargs)
-    else:
-        out = df.applymap(func, na_action=na_action, **kwargs)
-    return out
